+++ test/test_cpython_info.py
@@ -604,9 +604,11 @@ def test_is_released_unknown(pyvinfo: CPythonVersionInfo, v: str) -> None:
 
 
 @pytest.mark.parametrize(
-    "series,eol",
+    "version,eol",
     [
+        ("0", None),
         ("0.9", None),
+        ("1", None),
         ("1.0", None),
         ("1.1", None),
         ("1.2", None),
@@ -614,6 +616,7 @@ def test_is_released_unknown(pyvinfo: CPythonVersionInfo, v: str) -> None:
         ("1.4", None),
         ("1.5", None),
         ("1.6", None),
+        ("2", None),
         ("2.0", None),
         ("2.1", None),
         ("2.2", None),
@@ -621,7 +624,10 @@ def test_is_released_unknown(pyvinfo: CPythonVersionInfo, v: str) -> None:
         ("2.4", None),
         ("2.5", None),
         ("2.6", date(2013, 10, 29)),
+        ("2.6.9", date(2013, 10, 29)),
         ("2.7", date(2020, 1, 1)),
+        ("2.7.0", date(2020, 1, 1)),
+        ("3", None),
         ("3.0", date(2009, 1, 13)),
         ("3.1", date(2012, 6, 1)),
         ("3.2", date(2016, 2, 20)),
@@ -631,22 +637,21 @@ def test_is_released_unknown(pyvinfo: CPythonVersionInfo, v: str) -> None:
         ("3.6", date(2021, 12, 23)),
         ("3.7", date(2023, 6, 27)),
         ("3.8", date(2024, 10, 1)),
+        ("4", None),
         ("4.0", None),
     ],
 )
 def test_eol_date(
-    pyvinfo: CPythonVersionInfo, series: str, eol: Optional[date]
+    pyvinfo: CPythonVersionInfo, version: str, eol: Optional[date]
 ) -> None:
-    assert pyvinfo.eol_date(series) == eol
+    assert pyvinfo.eol_date(version) == eol
 
 
-@pytest.mark.parametrize(
-    "v", INVALID_VERSIONS + ["2.5.7", "3", "3.0.0", "3.0.1", "3.9.0", "5"]
-)
+@pytest.mark.parametrize("v", INVALID_VERSIONS)
 def test_eol_date_invalid(pyvinfo: CPythonVersionInfo, v: str) -> None:
     with pytest.raises(ValueError) as excinfo:
         pyvinfo.eol_date(v)
-    assert str(excinfo.value) == f"Invalid series name: {v!r}"
+    assert str(excinfo.value) == f"Invalid version string: {v!r}"
 
 
 @pytest.mark.parametrize("v", ["0.8", "3.9"])
@@ -658,9 +663,11 @@ def test_eol_date_unknown(pyvinfo: CPythonVersionInfo, v: str) -> None:
 
 
 @pytest.mark.parametrize(
-    "series,is_eol",
+    "version,is_eol",
     [
+        ("0", True),
         ("0.9", True),
+        ("1", True),
         ("1.0", True),
         ("1.1", True),
         ("1.2", True),
@@ -668,6 +675,7 @@ def test_eol_date_unknown(pyvinfo: CPythonVersionInfo, v: str) -> None:
         ("1.4", True),
         ("1.5", True),
         ("1.6", True),
+        ("2", False),
         ("2.0", True),
         ("2.1", True),
         ("2.2", True),
@@ -675,7 +683,10 @@ def test_eol_date_unknown(pyvinfo: CPythonVersionInfo, v: str) -> None:
         ("2.4", True),
         ("2.5", True),
         ("2.6", True),
+        ("2.6.9", True),
         ("2.7", False),
+        ("2.7.0", False),
+        ("3", False),
         ("3.0", True),
         ("3.1", True),
         ("3.2", True),
@@ -685,20 +696,19 @@ def test_eol_date_unknown(pyvinfo: CPythonVersionInfo, v: str) -> None:
         ("3.6", False),
         ("3.7", False),
         ("3.8", False),
+        ("4", False),
         ("4.0", False),
     ],
 )
-def test_is_eol(pyvinfo: CPythonVersionInfo, series: str, is_eol: bool) -> None:
-    assert pyvinfo.is_eol(series) is is_eol
+def test_is_eol(pyvinfo: CPythonVersionInfo, version: str, is_eol: bool) -> None:
+    assert pyvinfo.is_eol(version) is is_eol
 
 
-@pytest.mark.parametrize(
-    "v", INVALID_VERSIONS + ["2.5.7", "3", "3.0.0", "3.0.1", "3.9.0", "5"]
-)
+@pytest.mark.parametrize("v", INVALID_VERSIONS)
 def test_is_eol_invalid(pyvinfo: CPythonVersionInfo, v: str) -> None:
     with pytest.raises(ValueError) as excinfo:
         pyvinfo.is_eol(v)
-    assert str(excinfo.value) == f"Invalid series name: {v!r}"
+    assert str(excinfo.value) == f"Invalid version string: {v!r}"
 
 
 @pytest.mark.parametrize("v", ["0.8", "3.9"])