+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v1.1.0 (in development)
 -----------------------
 - Use pydantic internally for parsing & validating version databases
+- `eol_date()` and `is_eol()` now accept major and micro versions
 
 v1.0.0 (2021-11-04)
 -------------------