+++ docs/changelog.rst
@@ -6,6 +6,8 @@ Changelog
 v1.1.0 (in development)
 -----------------------
 - Use pydantic internally for parsing & validating version databases
+- `~CPythonVersionInfo.eol_date()` and `~CPythonVersionInfo.is_eol()` now
+  accept major and micro versions
 
 
 v1.0.0 (2021-11-04)